Added modified to-do list. Still trying to get it to add another note.

+++ website/auth.py
@@ -79,6 +79,7 @@ def logout():
 
 
 
+#experimenting on how to add another note
 @auth.route('/',methods=["GET", "[POST]"])
 def home():
     if request.method == '[POST]':
@@ -87,7 +88,7 @@ def home():
         if len(note) < 1:
                 flash('Note is too short!', category='error') 
         else:
-                new_note = Note(data=note)  #providing the schema for the note 
+                new_note = Note(data=note, user_id = current_user.id)  #providing the schema for the note 
                 db.session.add(new_note) #adding the note to the database 
                 db.session.commit()
                 flash('Note added!', category='success')
@@ -101,7 +102,8 @@ def delete_note():
     noteId = note['noteId']
     note = Note.query.get(noteId)
     if note:
-        db.session.delete(note)
-        db.session.commit()
+        if note.user_id == current_user.id:
+            db.session.delete(note)
+            db.session.commit()
 
     return jsonify({})
